Fix db export (#81)
* DBのフォーマットを見直した
ついでにreadmeを更新して未使用のパッケージを削除した

* 色々修正

* 色々改善

* ・装備可能条件を修正
・製品/建造リソース計算処理の改良
・総合保管庫のレコードを暫定的に保管庫タブに表示するようにした

* 兵装編集画面を動作するようにした

* 装備と建築計画からインポートする機能を動作するようにした

* DBビュワーを動作するようにした

* SQLを使用する箇所についてマッピングを使用するようにした

* 保存した.x4ファイルの読み込みが遅いのを修正

* ・一部の艦船のデータを抽出できないのを修正
・WareEquipmentテーブルのキーにGroupNameを追加(WareID, ConnectionNameが同じでGroupNameのみ異なる場合があるため)
・兵装編集画面の装備済み個数が誤って表示される場合があるのを修正

* NuGetパッケージの更新

* ・App.config.json がVerUpで上書きされないようにした
・App.config.json の内容をプロパティでラップするようにした

* 起動時に更新を確認するか選択できるようにした

* 起動時の更新確認のローカライズ対応

* DDSファイルがgzipで圧縮されていない場合を考慮するようにした

* 一部のパターンで言語解決出来ていない不具合を修正

* DB更新画面に詳細な進捗を表すプログレスバーを追加した

* ・DB抽出が遅いのを修正
・SizeテーブルのNameが言語フィールド文字列で抽出されるのを修正

* DBのフォーマットバージョンを更新

+++ X4_DataExporterWPF/Entity/Engine.cs
@@ -18,19 +18,19 @@ namespace X4_DataExporterWPF.Entity
         /// <summary>
         /// 前方推進力
         /// </summary>
-        public int ForwardThrust { get; }
+        public double ForwardThrust { get; }
 
 
         /// <summary>
         /// 後方推進力
         /// </summary>
-        public int ReverseThrust { get; }
+        public double ReverseThrust { get; }
 
 
         /// <summary>
         /// ブースト推進力
         /// </summary>
-        public int BoostThrust { get; }
+        public double BoostThrust { get; }
 
 
         /// <summary>
@@ -48,7 +48,7 @@ namespace X4_DataExporterWPF.Entity
         /// <summary>
         /// トラベル推進力
         /// </summary>
-        public int TravelThrust { get; }
+        public double TravelThrust { get; }
 
 
         /// <summary>
@@ -68,7 +68,7 @@ namespace X4_DataExporterWPF.Entity
         /// <param name="boostReleaseTime">ブースト解除時間</param>
         /// <param name="travelThrust">トラベル推進力</param>
         /// <param name="travelReleaseTime">トラベル解除時間</param>
-        public Engine(string equipmentID, int forwardThrust, int reverseThrust, int boostThrust, double boostDuration, double boostReleaseTime, int travelThrust, double travelReleaseTime)
+        public Engine(string equipmentID, double forwardThrust, double reverseThrust, double boostThrust, double boostDuration, double boostReleaseTime, double travelThrust, double travelReleaseTime)
         {
             EquipmentID = equipmentID;
             ForwardThrust = forwardThrust;
